Get comments

+++ ServerApp/Interfaces/CommentService.cs
@@ -7,13 +7,13 @@ namespace ServerApp.Interfaces
     public interface ICommentService
     {
         Task<bool> AddAsync(Comment comment);
-        Task<bool> AddDislikeAsync(int commentId, long userId);
-        Task<bool> AddLikeAsync(int commentId, long userId);
-        Task<bool> ApproveAsync(int commentId);
+        Task<bool> AddDislikeAsync(long commentId, long userId);
+        Task<bool> AddLikeAsync(long commentId, long userId);
+        Task<bool> ApproveAsync(long commentId);
         Task<List<Comment>> GetAsync(long userId);
         Task<List<Comment>> GetByArticleAsync(long articleId);
-        Task<bool> RejectAsync(int commentId);
-        Task<bool> RemoveDislikeAsync(int commentId, long userId);
-        Task<bool> RemoveLikeAsync(int commentId, long userId);
+        Task<bool> RejectAsync(long commentId);
+        Task<bool> RemoveDislikeAsync(long commentId, long userId);
+        Task<bool> RemoveLikeAsync(long commentId, long userId);
     }
 }
